% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{convert_to_long}
\alias{convert_to_long}
\title{convert_to_long}
\usage{
convert_to_long(data, extra_col)
}
\arguments{
\item{data}{dataset with cases and country data}

\item{extra_col}{columns for country data}
}
\value{
dataset converted to long
}
\description{
converts dataset with dates to long, removes
}
