% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_forecasting.R
\name{create_country_models}
\alias{create_country_models}
\title{Title}
\usage{
create_country_models(country_data, n)
}
\arguments{
\item{country_data}{data with countries to forecast on}

\item{n}{forecast horizon}
}
\value{
modeltable with models trained on country data
}
\description{
Title
}
