% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{create_daily_cases}
\alias{create_daily_cases}
\title{create_daily_cases}
\usage{
create_daily_cases(data)
}
\arguments{
\item{data}{dataframe with converted data}
}
\value{
daily cases
}
\description{
creates columns with daily case numbers
}
