% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{create_graph_non_forecast}
\alias{create_graph_non_forecast}
\title{Title}
\usage{
create_graph_non_forecast(
  graph_choice,
  country_list,
  date_range_start,
  date_range_end,
  data_value,
  title,
  y_lab
)
}
\arguments{
\item{graph_choice}{choice of graph}

\item{country_list}{countries chosen}

\item{date_range_start}{date start}

\item{date_range_end}{date end}

\item{data_value}{column with data}

\item{title}{title}

\item{y_lab}{y axis label}
}
\value{
graphs without forecasts
}
\description{
Title
}
