% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{create_labels}
\alias{create_labels}
\title{create_labels}
\usage{
create_labels(country_list)
}
\arguments{
\item{country_list}{list of countries input}
}
\value{
list of countries and flags appended
}
\description{
creates labels for plot legend with coutry names and flags
}
