% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{data_prep}
\alias{data_prep}
\title{data_prep}
\usage{
data_prep(data, country_data)
}
\arguments{
\item{data}{dataset containing case data}

\item{country_data}{dataset containing country data}
}
\value{
merged dataset prepared for plotting
}
\description{
prepares country data for use in dataset
}
