% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{data_prep_wrapper}
\alias{data_prep_wrapper}
\title{data_prep_wrapper}
\usage{
data_prep_wrapper(url_confirmed, url_deaths)
}
\arguments{
\item{url_confirmed}{url to confirmed cases}

\item{url_deaths}{url to death cases}
}
\value{
dataset containing cases and country data
}
\description{
wrapper function for data prep
}
