% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_forecasting.R
\name{forecast_on_country_premodelled}
\alias{forecast_on_country_premodelled}
\title{forecasting function TBD}
\usage{
forecast_on_country_premodelled(
  data,
  country,
  date_range_start,
  date_range_end,
  models,
  forecast_horizon,
  calibration_table
)
}
\arguments{
\item{data}{forecasting data}

\item{country}{countries to forecast on}

\item{date_range_start}{datestart}

\item{date_range_end}{dateend}

\item{models}{models to use}

\item{forecast_horizon}{forecast horizon}

\item{calibration_table}{list with all countries modelled}
}
\value{
forecast plot on premodelled data
}
\description{
forecasting function TBD
}
