% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_html}
\alias{get_html}
\title{get_html}
\usage{
get_html(webpage, x)
}
\arguments{
\item{webpage}{John Hopkins scrape site}

\item{x}{child node to scrape}
}
\value{
data from node
}
\description{
scrapes nodes from online datatable
}
