% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{renderPlot_function}
\alias{renderPlot_function}
\title{renderplot function}
\usage{
renderPlot_function(input, plot_type)
}
\arguments{
\item{input}{input, userbased}

\item{plot_type}{type of plot, case_number or cases_pr_citizen}
}
\value{
graph
}
\description{
renders graphs
}
