% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{server function}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{inputs for server}

\item{output}{output to server UI}

\item{session}{session data}
}
\value{
server output
}
\description{
server function
}
